refactor: unify ignore-pattern glob matching into mdbook-lint-core
The CLI/preprocessor `path_is_ignored` and MDBOOK005's `matches_ignore_pattern`
were line-for-line duplicates of the same .gitignore-style glob matcher. Move
the single implementation into mdbook-lint-core::ignore so ignore-paths and
MDBOOK005's ignore_patterns route through it and can no longer drift apart.

mdbook-lint-cli::ignore now re-exports the shared path_is_ignored and keeps
filter_ignored_paths as a thin wrapper. mdbook005.rs calls the shared
function instead of its own copy; its duplicate unit test is dropped in
favor of the matching-semantics tests that now live with the shared
implementation.

Fixes #479
