+++ crates/mdbook-lint-cli/Cargo.toml
@@ -37,7 +37,6 @@ comrak = { workspace = true }
 clap = { workspace = true }
 mdbook = { workspace = true }
 walkdir = { workspace = true }
-glob = { workspace = true }
 rayon = "1.10"
 tabled = "0.20"
 anstream = "0.6"
